==> R/00_all.qmd <==
---
title: "Gene expression analysis of breast cancer proteome"
format:
  html:
    embed-resources: true
editor: visual
---

```{r}
# Run all documents

#| label: 00_Load packages
#| message: false
#| output: false
library(quarto)
```

```{r}
###### FILE 01
quarto::quarto_render("01_load_DONE.qmd",
                      execute = T)
file.rename(from = "01_load_DONE.html",
            to = "../results/01_load_DONE.html")
file.rename(from = "01_load_DONE_files/",
            to = "../results/01_load_DONE_files")


###### FILE 02
quarto::quarto_render("02_clean_DONE.qmd")
file.rename(from = "02_clean_DONE.html",
            to = "../results/02_clean_DONE.html")
file.rename(from = "02_clean_DONE_files/",
            to = "../results/02_clean_DONE_files")


###### FILE 03
quarto::quarto_render("03_augment_DONE.qmd",
                      output_format = "html", 
                      execute = T)
file.rename(from = "03_augment_DONE.html",
            to = "../results/03_augment_DONE.html")
file.rename(from = "03_augment_DONE_files/",
            to = "../results/03_augment_DONE_files")


###### FILE 04
quarto::quarto_render("04_describe_DONE.qmd",
                      output_format = "html",
                      execute = T)
file.rename(from = "04_describe_DONE.html",
            to = "../results/04_describe_DONE.html")
file.rename(from = "04_describe_DONE_files/",
            to = "../results/04_describe_DONE_files")


###### FILE 05
quarto::quarto_render("05_analysis_clustering_DONE.qmd",
                      output_format = "html",
                      execute = T)
file.rename(from = "05_analysis_clustering_DONE.html",
            to = "../results/05_analysis_clustering_DONE.html")
file.rename(from = "05_analysis_clustering_DONE_files/",
            to = "../results/05_analysis_clustering_DONE_files")


###### FILE 06
quarto::quarto_render("06_analysis_PCA_cluster_DONE.qmd",
                      output_format = "html",
                      execute = T)
file.rename(from = "06_analysis_PCA_cluster_DONE.html",
            to = "../results/06_analysis_PCA_cluster_DONE.html")
file.rename(from = "06_analysis_PCA_cluster_DONE_files",
            to = "../results/06_analysis_PCA_cluster_DONE_files")

###### FILE 07
quarto::quarto_render("07_analysis_diff_exp_DONE.qmd.qmd",
                      output_format = "html",
                      execute = T)
file.rename(from = "07_analysis_diff_exp_DONE.html",
            to = "../results/07_analysis_diff_exp_DONE.html")
file.rename(from = "07_analysis_diff_exp_DONE_files/",
            to = "../results/07_analysis_diff_exp_DONE_files")

###### FILE 08
quarto::quarto_render("08_analysis_kmeans_DONE.qmd",
                      output_format = "html",
                      execute = T)
file.rename(from = "08_analysis_kmeans_DONE.html",
            to = "../results/08_analysis_kmeans_DONE.qmd.html")
file.rename(from = "08_analysis_kmeans_DONE_files/",
            to = "../results/08_analysis_kmeans_DONE_files")

###### FILE 09
quarto::quarto_render("09_analysis_volcano_plots_DONE.qmd",
                      output_format = "html",
                      execute = T)
file.rename(from = "09_analysis_volcano_plots_DONE.html",
            to = "../results/09_analysis_volcano_plots_DONE.html")
file.rename(from = "09_analysis_volcano_plots_DONE_files/",
            to = "../results/09_analysis_volcano_plots_DONE_files")


###### FILE 10
quarto::quarto_render("10_analysis_heatmaps_diff_exp_DONE.qmd",
                      output_format = "html",
                      execute = T)
file.rename(from = "10_analysis_heatmaps_diff_exp_DONE.html",
            to = "../results/10_analysis_heatmaps_diff_exp_DONE.html")
file.rename(from = "10_analysis_heatmaps_diff_exp_DONE_files/",
            to = "../results/10_analysis_heatmaps_diff_exp_DONE_files")


###### FILE 11
quarto::quarto_render("11_analysis_dendro_signif_proteins_DONE.qmd",
                      output_format = "html",
                      execute = T)
file.rename(from = "11_analysis_dendro_signif_proteins_DONE.html",
            to = "../results/11_analysis_dendro_signif_proteins_DONE.html")
file.rename(from = "11_analysis_dendro_signif_proteins_DONE_files/",
            to = "../results/11_analysis_dendro_signif_proteins_DONE_files")

###### FILE 12
quarto::quarto_render("12_correlation_DONE.qmd",
                      output_format = "html",
                      execute = T)
file.rename(from = "12_correlation_DONE.html",
            to = "../results/12_correlation_DONE.html")
file.rename(from = "12_correlation_DONE_files/",
            to = "../results/12_correlation_DONE_files")


```

## Include all scripts in 00_all.html

{{< include 01_load_DONE.qmd >}}

{{< include 02_clean_DONE.qmd >}}

{{< include 03_augment_DONE.qmd >}}

{{< include 04_describe_DONE.qmd >}}

{{< include 05_analysis_clustering_DONE.qmd >}}

{{< include 06_analysis_PCA_cluster_DONE.qmd >}}

{{< include 07_analysis_diff_exp_DONE.qmd >}}

{{< include 08_analysis_kmeans_DONE.qmd >}}

{{< include 09_analysis_volcano_plots_DONE.qmd >}}

{{< include 10_analysis_heatmaps_diff_exp_DONE.qmd >}}

{{< include 11_analysis_dendro_signif_proteins_DONE.qmd >}}

{{< include 12_correlation_DONE.qmd >}}

```{r}

```
