==> R/01_load_DONE.qmd <==
---
title: "01 loading of data"
format:
  html:
    toc: true
    html-math-method: katex
    css: styles.css
---

```{r}
#| label: load packages and data
rm(list = ls())
library(curl)
library(tidyverse)
library("quarto")
protein_data <- read_csv2("../_raw/77_cancer_proteomes_CPTAC_itraq.csv")
clinical_data <- read_csv("../_raw/clinical_data_breast_cancer.csv")
PAM50_file <- read_csv("../_raw/PAM50_proteins.csv")
```

```{r}
#| label: write data to csv file
#write_csv2 converts the count matrix to a variable of characters
write_csv2(protein_data, "../data/protein_load.csv")
write_csv2(clinical_data, "../data/clinical_load.csv")
write_csv2(PAM50_file, "../data/PAM50_load.csv")
```
