==> R/05_analysis_1_clustering_DONE.qmd <==
```{r}

```

---
title: "06_hc_dendro"
format: html
editor: visual
---

# Clustering analysis with all proteomics data

## Hierachial clustering visualised with dendrograms

To investigate if the clinically assigned pam50 categories display in the proteomics dataset, we also wanted to do a hierachial clustering analysis. This was visualised using dendrograms.

First, packages and data were loaded.

```{r}
#| echo: false
#| label: load packages and data

# Packages
library(tidyverse)

if (!require('ggendro', character.only = TRUE)) {
# If not installed, install the package
 install.packages('ggdendro')
}
library(ggdendro)

# data
clinical_data <- read.table("../data/clinical_data.tsv", sep = "\t", header = T)
cm <- read.table("../data/count_matrix_1_77.tsv", header = T) |>
  pivot_longer(cols = -id, names_to = "id2", values_to = "value")|> # Transforming to pivot_longer
  pivot_wider(names_from = id, values_from = value) |>
  column_to_rownames("id2")

```

### Analysis based on entire proteomics data set

Next, the hierachial clustering was done.

```{r}
#| label: hierachial clustering

# create distance matrix (scaled) and perform hierachial clustering
hc <- cm |>
  stats::dist() |> # calculating the distance for each value
  stats::hclust(method = "ward.D2")  # applying the hierachial clustering algorithm to the data 

# save clustering as dendogram
ddg <- stats::as.dendrogram(hc)
  
# extract dendogram order
ddg_order <- stats::order.dendrogram(ddg)
  
# extract all dendro information
ddg_data <- ggdendro::dendro_data(ddg)

# add pam50 information to ddg_data
ddg_label <- ggdendro::label(ddg_data) |>
  dplyr::rename(id = label) |> 
  dplyr::left_join(clinical_data, by = c("id")) |> 
  dplyr::select(x,y, id, pam50)

```

This was followed by visualisation in the form of a dendrogram using the ggdendro package.

```{r}
#| label: dendrogram

# create dendogram plot based on hierachial clustering
ddg_plot <- ggdendro::ggdendrogram(data = hc, rotate=TRUE) +
  theme(axis.text.y = element_blank(), # removing labels on x axis
        axis.text.x = element_blank(), # removing labels on Y axis
        plot.title = element_text(hjust = 0.5, vjust = 0)) +  
  geom_text(data=ddg_label,
            aes(label=id, x=x, y=-45, colour=pam50), # labels on y axis colored by pam50. 
            size = 1.5) +
  labs(title = "Hierachial clustering based on full proteomics dataset")

ddg_plot

# save dendrogram plot
ggsave("../results/hc_dendro_all.png", ddg_plot, height = 4, width = 7)
```

**Conclusion**: From the colored labels, it is apparent, that the samples don't group by pam50 categories when a hierachial clustering is made on the entire proteomics dataset.

### Analysis based on significant proteins

We want to see if we get a better clustering when limiting the hierachial analysis to only consider the genes that were significant in our differential expression analysis.

A count matrix with a reduced dataset - with the proteomics data of only the significant proteins - were loaded.

```{r}
#| label: load data
cm_sig <- read.table("../data/count_matrix_significant_proteins.tsv", header = T) |>
  column_to_rownames('id')
```

Next, the hierachial clustering was done exactly as above.

```{r}
#| label: hierachial clustering 2

# create distance matrix (scaled) and perform hierachial clustering
hc <- cm_sig |>
  dist() |> # calculating the distance for each value
  hclust(method = "ward.D2")  # applying the hierachial clustering algorithm to the data 

# save clustering as dendogram
ddg <- as.dendrogram(hc)
  
# extract dendogram order
ddg_order <- order.dendrogram(ddg) # delete if deleting heatmap
  
# extract all dendro information
ddg_data <- dendro_data(ddg)

# add pam50 information to ddg_data
ddg_label <- label(ddg_data) |>
  rename(id = label) |> 
  left_join(clinical_data, by = c("id")) |> 
  select(x,y, id, pam50)

```

And the dendrogram visualised as above.

```{r}

#| label: dendrogram 2

# create dendogram plot based on hierachial clustering
ddg_plot2 <- ggdendrogram(data = hc, rotate=TRUE) +
  theme(axis.text.y = element_blank(), # removing labels on x axis
        axis.text.x = element_blank(), # removing labels on Y axis
        plot.title = element_text(hjust = 0.5, vjust = 0)) +  
  geom_text(data=ddg_label,
            aes(label=id, x=x, y=-15, colour=pam50), # labels on y axis colored by pam50. 
            size = 1.5) +
  labs(title = "Hierachial clustering based on significant proteins")

ddg_plot2

# save dendrogram plot
ggsave("../results/hc_dendro_sig.png", ddg_plot2, height = 4, width = 7)
```

**Conclusion**: After filtering the dataset on only the significant proteins based on our differential expression analysis, we see a much better agreement between the hierachial clustering of proteomics data and the clinical pam50 assignments.

### Alternative method for visualisation TO BE REMOVED

By using the pheatmap package, we can couple the hierachial clustering analysis and visualisation in a dendogram (as done above) with that in a heatmap in very few steps of code:

```{r}
library(pheatmap)

cm2 <- read.table("../data/count_matrix_1_77.tsv", header = T) |>
  column_to_rownames('id')

pheatmap(
  mat = cm2,
  clustering_method = "ward.D2", # use hierachial clustering method
  scale = "row",
  show_rownames = FALSE,
  main = "Hierachial clustering analysis, all proteomics data",
  treeheight_row = 0, #do not display row tree
  fontsize_col= 3
  )
```

\
\
\
\*\*\* I suggest to delete below because I can't figure out how to order the rows like in the dendogram \*\*\*\

```{r}
#| label: visualization

# define long data structure based on the count matrix
cm_long <- cm |> 
  rownames_to_column(var = "id") |> # setting rownames to a column named id
  pivot_longer(cols = -c("id"), names_to = "protein") |> # converting to pivot longer
  mutate(id = factor(id, levels = unique(id))) # Factorize the id column based on unique values from id ordered by the ddg_order

# heatmap on scaled data:
hm <- ggplot(data = cm_long, mapping = aes(x = protein, y = value)) +
        geom_tile(aes(fill=value)) + # heatmap function
        scale_fill_gradient2(low = "red", high = "blue") + # scale color (high values are blue, low values are red)
        theme(axis.text.x = element_blank(), # remove label on x axis
              axis.text.y = element_blank(),
              legend.position = "left") +
        geom_text(data=ddg_label,
            aes(label=id, x=x, y=-20 ,colour=pam50), # labels on y axis colored by pam50. 
            size = 1.5) + 
  theme_minimal() + # position of the legend on the left
hm


```

Print the two plots side-by-side using patchwork:

```{r}
library(patchwork)
hm + ddg_plot 
```

```{r}
rm(list = ls())
```
