==> R/06_analysis_PCA_cluster.qmd <==
---
title: "05_PCA"
format: html
editor: visual
---

# Clustering analysis with all proteomics data

## Principal component analysis (PCA)

It was investigated if the data set would show any clustering resembling that of the PAM50 criteria, based on the entire proteomics data set. First a PCA analysis was conducted.

First, required packages were loaded.

```{r}
#| echo: false
#| label: load packages

library(tidyverse)
library(broom)
library(cowplot) # only used for some aesthetics

```

Next, the required data was loaded. For this analysis it was the count matrix, which was transformed (rows to columns and vice versa), and the clinical data.

```{r}
#| label: load data (and augment)

cm <- read.table("../data/count_matrix_1_77.tsv", header = T) |>
  pivot_longer(cols = -id, names_to = 'variable', values_to = 'value') |>
  pivot_wider(names_from = id, values_from = value) |>
  rename(id = variable) |>
  column_to_rownames('id')

clinical_data <- read.table("../data/clinical_data.tsv", sep = "\t", header = T)
```

Next, the PCA was done by first scaling the count matrix, and then the doing the actual principal component analysis using the prcomp() function. From the pca output, information about the variance-explanation by each principel component was extracted and stored in the pca_results using the tidy() function.

```{r}
#| label: define PCA

pca <- cm |> 
  stats::prcomp(center=TRUE, scale=TRUE) # transform data to principal components

#Extract variance explained by the PCs
pca_results <- pca |>
  broom::tidy("pcs") |> # Using broom::tidy to extract results
  purrr::pluck("percent") # extracting the explained varaince of each PC

  
```

These variables, were used for creating a visual presentation of the principal component analysis, in the form of a classic PCA plot of the first to principal components. Here the individual samples were coloured based on their pam50 classification.

```{r}
#| label: PCA plot

# Create PCA plot
pca_plot <- pca |> 
  pluck("x") |> 
  data.frame() |> 
   ggplot(aes(x=PC1, 
              y=PC2, 
              color = clinical_data |> dplyr::pull("pam50"), # color by pam50 from clinical_data
              fill = clinical_data |> dplyr::pull("pam50"))) +
   geom_point(shape = 21,
             size = 3, 
             alpha = 0.5) +
   theme_bw(base_size = 12) + # font size of text
   background_grid() +
   xlab(str_c('PC1,  ', round(pca_results[1]*100, digits=1), '% VE')) + 
   ylab(str_c('PC2,  ', round(pca_results[2]*100, digits=1), '% VE')) +
   labs(
     title = "PCA on entire proteomics dataset",
     color = "pam50",
     fill = "pam50"
        )
#print the plot
pca_plot

#save the plot
ggsave("../results/PCA1.png", pca_plot, height = 4, width = 6)

```

In this PCA based on all the proteomics data, we see no obvious clustering.

```{r}
#quarto::quarto_render("06_analysis_PCA_cluster.qmd",output_format = "html",execute = T)
#file.rename(from = "06_analysis_PCA_cluster.html",
#            to = "../results/06_analysis_PCA_cluster.html")
```

```{r}
rm(list = ls())
```
