==> R/08_analysis_volcano_plots_DONE.qmd <==
---
title: "Differential expression analysis: Volcano plot"
format:
  html:
    toc: true
    html-math-method: katex
    css: styles.css
---

```{r}
#| label: loading packages
library(tidyverse)
library(patchwork)
library(ggrepel)
library(patchwork)
source("99_proj_func.R")

#read files.
HER2 <- read_table("../data/diff_exp_HER2.tsv")
lumA <- read_table("../data/diff_exp_Luminal_A.tsv")
lumB <- read_table("../data/diff_exp_Luminal_B.tsv")
basal_like <- read_table("../data/diff_exp_Basal_like.tsv")
```

To see dispersion of differentially expressed genes, we visualize then through a volcanoplot. to make the volcanoplot, we call the function `volcano_plot()` from the `99_proj_func.R` file.

```{r}
#| label: analysis
plot1 <- volcano_plot(data = HER2, subtype = "HER2")
plot2 <- volcano_plot(data = lumA, subtype = "Luminal A")
plot3 <- volcano_plot(data = lumB, subtype = "Luminal A")
plot4 <- volcano_plot(data = basal_like, subtype = "Basal like")
```

```{r}
#| label: save plots
#loads patchwork library to combine the four volcanoplots where / creates to row #and two columns, and collect ensures that legends are shared.
#the file with combined plot is saved as png file in "/results/ directory with specified dimentions.
volcano_plots <- (plot1 + plot2)/(plot3 + plot4) +
  plot_layout(guides = "collect")
ggsave("../results/volcano_plots.png", volcano_plots, height = 8, width = 16, dpi = 300)

```

```{r}
#| label: quarto rendering
#quarto::quarto_render("08_analysis_volcano_plot_DONE.qmd",output_format = "html",execute = T)
#file.rename(from = "08_analysis_volcano_plot_DONE.html",
#            to = "../results/08_analysis_volcano_plot_DONE.html")
```
