==> R/10_analysis_heatmaps_diff_exp_DONE.qmd <==
---
title: "Differential expression analysis: Heatmap"
format:
  html:
    toc: true
    html-math-method: katex
    css: styles.css
---

```{r}
#| label: loading packages and data
library(tidyverse)
library(patchwork)
source("99_proj_func.R")

HER2 <- read.table("../data/diff_exp_HER2.tsv", header = T, sep = "\t") |>
  subset_rename_data(subtype = "HER2")

lumA <- read.table("../data/diff_exp_Luminal_A.tsv", header = T, sep = "\t")|>
  subset_rename_data(subtype = "lumA")

lumB <- read.table("../data/diff_exp_Luminal_B.tsv", header = T, sep = "\t")|>
  subset_rename_data(subtype = "lumB")

basal_like <- read.table("../data/diff_exp_Basal_like.tsv", header = T, sep = "\t")|>
  subset_rename_data(subtype = "bl")

```

**Load data**

In the following, the results from the differential gene expression analysis is loaded and wrangled by ise of a function `subset_rename_data` from `99_proj_func.R` . The function subsets the data to only the columns of interest for this analysis and renames the coloumns.

```{r}
#| label: analysis
#Extract only neccesary columns and rename for further joining: 

# Find only the significant proteins:
signif_fdr_prots <- c(HER2 |> filter(adjustedfdr_HER2 <= 0.05) |> pull(protein),
                  lumA |> filter(adjustedfdr_lumA <= 0.05) |> pull(protein), 
                  lumB |> filter(adjustedfdr_lumB <= 0.05) |> pull(protein), 
                  basal_like |> filter(adjustedfdr_bl <= 0.05) |> pull(protein))
    # Make a vector of the significant proteins

signif_fdr_prots_unique <- unique(signif_fdr_prots)
    # Select eah protein inly once.  
rm(signif_fdr_prots)
```

Combine the datasets to the dataframe we will use:

```{r}
#| label: analysis 2
combined <- left_join(HER2, lumA, by = "protein") |>
      #Combine results dataframes
  left_join(lumB, by = "protein") |>
      #Combine results dataframes
  left_join(basal_like, by = "protein") |>
      #Combine results dataframes
  filter(protein %in% signif_fdr_prots_unique)
      # Filter for only the significant protiens

rm(basal_like, lumA, lumB, HER2, signif_fdr_prots_unique)
      # Remove unneccesary stuff from enviroenment
```

Next, the heatmaps of the results are made using the `heat_map_plot` function from the plots file. Because we had so many significant proteins, a visualization of them all in the plot was un-informational. Therefore, we chose to only select the top 20 most significant proteins for each subgroup to display. In summary, this function...

-   Takes the combined data and order it in order of significance for subtype of choice.

-   Structures the data, making sure, the proteins are plotted in order of significance of subtype in question.

-   Makes a heatmap colouring for estimate and indicating significance by the use of asterisks.

```{r}
#| label: analysis 3
heatmap_bl <- heat_map_results(data = combined, fdr_col = adjustedfdr_bl, title = "Basal like") + theme(legend.position = 'none')

heatmap_HER2 <- heat_map_results(data = combined, fdr_col = adjustedfdr_HER2, title = "HER2")

heatmap_lumA <- heat_map_results(data = combined, fdr_col = adjustedfdr_lumA, title = "Luminal A")+ theme(legend.position = 'none')

heatmap_lumB <- heat_map_results(data = combined, fdr_col = adjustedfdr_lumB, title = "Luminal B")+ theme(legend.position = 'none')

```

```{r}
#| label: save results
#Combine all heatmaps using 'patchwork'
heatmap <- (heatmap_bl + heatmap_HER2)/(heatmap_lumA +heatmap_lumB)+
  plot_layout(guides = 'collect') +
  plot_annotation(title = "Vizualisation of results from differential gene expression analysis", 
        subtitle = "Using top 20 most significant proteins for each PAM50 classification")
ggsave("../results/heatmap_combined.png", heatmap, width = 12, height = 10, dpi = 300)
#Save
```

```{r}
rm(list = ls())
```
